Trait pallet_xcm_bridge_hub::pallet::Config

pub trait Config<I: 'static = ()>: BridgeMessagesConfig<Self::BridgeMessagesPalletInstance> {
    type UniversalLocation: Get<InteriorLocation>;
    type BridgedNetwork: Get<Location>;
    type BridgeMessagesPalletInstance: 'static;
    type MessageExportPrice: Get<Assets>;
    type DestinationVersion: GetVersion;
    type Lanes: Get<Vec<(SenderAndLane, (NetworkId, InteriorLocation))>>;
    type LanesSupport: XcmBlobHauler;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type UniversalLocation: Get<InteriorLocation>

Runtime’s universal location.

type BridgedNetwork: Get<Location>

Bridged network as relative location of bridged GlobalConsensus.

type BridgeMessagesPalletInstance: 'static

Associated messages pallet instance that bridges us with the BridgedNetworkId consensus.

type MessageExportPrice: Get<Assets>

Price of single message export to the bridged consensus (Self::BridgedNetworkId).

type DestinationVersion: GetVersion

Checks the XCM version for the destination.

type Lanes: Get<Vec<(SenderAndLane, (NetworkId, InteriorLocation))>>

Get point-to-point links with bridged consensus (Self::BridgedNetworkId). (this will be replaced with dynamic on-chain bridges - Bridges V2)

type LanesSupport: XcmBlobHauler

Support for point-to-point links (this will be replaced with dynamic on-chain bridges - Bridges V2)

Object Safety

This trait is not object safe.